feat(reviews): add Review entity, table, and EF configurations with constraints (#6)

+++ src/Engzly.Application/Features/Reviews/Commands/PostReview/PostReviewCommand.cs
@@ -7,5 +7,5 @@ public sealed record PostReviewCommand(
         string ReviewedUserId,
         string GigId,
         string? Comment,
-        float Rating
+        decimal Rating
     ) : IRequest<string>;
